image/test/mochitest/test_short_gif_header.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=844684
-->
<head>
  <title>Test for Bug 844684</title>
  <script type="application/javascript" src="/MochiKit/MochiKit.js"></script>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=844684">Mozilla Bug 844684</a>
<div id="content">
<img id="testcontent" onload="success()" onerror="failure()">
</div>
<pre id="test">
<script type="application/javascript">
SimpleTest.waitForExplicitFinish();

document.getElementById('testcontent').src = "short_header.gif";

function success() {
  ok(true, "Image loaded.");
  SimpleTest.finish();
}

function failure() {
  ok(false, "Image didn't load.");
  SimpleTest.finish();
}
</script>
</pre>
</body>
</html>
